Integración-patron-repository

+++ Web_Reserva_Vet/Controllers/UsersController.cs
@@ -6,8 +6,8 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using Web_Vet_Pet.Data;
-using Web_Vet_Pet.Models;
 using Web_Vet_Pet.Interfaces;
+using Web_Vet_Pet.Models;
 
 namespace Web_Vet_Pet.Controllers
 {
@@ -20,42 +20,34 @@ namespace Web_Vet_Pet.Controllers
             _userRepository = userRepository;
         }
 
-        // GET: Users
+        // GET: /Users
         public async Task<IActionResult> Index()
         {
             var users = await _userRepository.GetAllAsync();
             return View(users);
         }
 
-        // GET: Users/Details/5
-        public async Task<IActionResult> Details(int? id)
+        // GET: /Users/Details/5
+        public async Task<IActionResult> Details(int id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var user = await _userRepository.GetByIdAsync(id.Value);
+            var user = await _userRepository.GetByIdAsync(id);
             if (user == null)
             {
                 return NotFound();
             }
-
             return View(user);
         }
 
-        // GET: Users/Create
+        // GET: /Users/Create
         public IActionResult Create()
         {
             return View();
         }
 
-        // POST: Users/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        // POST: /Users/Create
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,FirstName,Email,PasswordHash,LastName,Phone,DateBirthday")] User user)
+        public async Task<IActionResult> Create(User user)
         {
             if (ModelState.IsValid)
             {
@@ -65,15 +57,10 @@ namespace Web_Vet_Pet.Controllers
             return View(user);
         }
 
-        // GET: Users/Edit/5
-        public async Task<IActionResult> Edit(int? id)
+        // GET: /Users/Edit/5
+        public async Task<IActionResult> Edit(int id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var user = await _userRepository.GetByIdAsync(id.Value);
+            var user = await _userRepository.GetByIdAsync(id);
             if (user == null)
             {
                 return NotFound();
@@ -81,57 +68,84 @@ namespace Web_Vet_Pet.Controllers
             return View(user);
         }
 
-        // POST: Users/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        // POST: /Users/Edit/5
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,FirstName,Email,PasswordHash,LastName,Phone,DateBirthday")] User user)
+        public async Task<IActionResult> Edit(int id, User user)
         {
             if (id != user.Id)
             {
-                return NotFound();
+                return BadRequest();
             }
-
             if (ModelState.IsValid)
             {
-                try
+                var existingUser = await _userRepository.GetByIdAsync(id);
+                if (existingUser == null)
                 {
-                    await _userRepository.UpdateAsync(user);
-                    return RedirectToAction(nameof(Index));
+                    return NotFound();
                 }
-                catch (DbUpdateConcurrencyException)
-                {
-                    return Problem("Error al actualizar el usuario.");
-                }
+
+                // Actualizar las propiedades de la entidad existente
+                existingUser.FirstName = user.FirstName;
+                existingUser.LastName = user.LastName;
+                existingUser.Email = user.Email;
+                existingUser.PasswordHash = user.PasswordHash; // Considera hashear la contraseña si cambió
+                existingUser.Phone = user.Phone;
+                existingUser.DateBirthday = user.DateBirthday;
+
+                await _userRepository.UpdateAsync(existingUser);
+                return RedirectToAction(nameof(Index));
             }
             return View(user);
         }
 
-        // GET: Users/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        // GET: /Users/Delete/5
+        public async Task<IActionResult> Delete(int id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var user = await _userRepository.GetByIdAsync(id.Value);
+            var user = await _userRepository.GetByIdAsync(id);
             if (user == null)
             {
                 return NotFound();
             }
-
             return View(user);
         }
 
-        // POST: Users/Delete/5
+        // POST: /Users/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            await _userRepository.DeleteAsync(id);
+            var user = await _userRepository.GetByIdAsync(id);
+            if (user != null)
+            {
+                await _userRepository.DeleteAsync(id);
+            }
             return RedirectToAction(nameof(Index));
+        }
+
+        // GET: /Users/ByEmail?email=example@email.com
+        public async Task<IActionResult> ByEmail(string email)
+        {
+            var user = await _userRepository.GetByEmailAsync(email);
+            if (user == null)
+            {
+                return NotFound();
+            }
+            return View("Details", user);
+        }
+
+        // GET: /Users/WithClients
+        public async Task<IActionResult> WithClients()
+        {
+            var users = await _userRepository.GetWithClientsAsync();
+            return View(users);
+        }
+
+        // GET: /Users/WithAdministrators
+        public async Task<IActionResult> WithAdministrators()
+        {
+            var users = await _userRepository.GetWithAdministratorsAsync();
+            return View(users);
         }
     }
 }
